# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Agadir Task Manager API
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: ["pending","done"] }
        due_date: { type: string, format: date-time }
paths:
  /auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        "201":
          description: Created
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
  /tasks:
    get:
      summary: List tasks
      security: [ { bearerAuth: [] } ]
      responses:
        "200": { description: OK }
    post:
      summary: Create task
      security: [ { bearerAuth: [] } ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        "201": { description: Created }
  /tasks/{id}:
    put:
      summary: Update task
      security: [ { bearerAuth: [] } ]
    delete:
      summary: Delete task
      security: [ { bearerAuth: [] } ]
  /tasks/{id}/done:
    patch:
      summary: Mark done
      security: [ { bearerAuth: [] } ]